requirements.txt
annotated-types==0.7.0
anyio==3.7.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
datasets==3.2.0
exceptiongroup==1.2.2
fastapi==0.109.2
filelock==3.17.0
fsspec<=2024.9.0,>=2023.1.0
google-api-core==2.24.1
google-api-python-client==2.118.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.29.1
idna==3.10
isodate==0.6.1
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4

packaging==24.2
pandas==2.1.3
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.1.0
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette>=0.36.3,<0.37.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.2.0
tqdm==4.67.1
transformers==4.37.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.27.1
youtube-transcript-api==0.6.2